rename cwtPayloadTypes -> cwtTypes
