+++ src/cwt.ts
@@ -3,7 +3,7 @@ import {
   UnvalidatedCWTClaims,
   UnvalidatedCWTHeaders,
   VC,
-} from "./cwtPayloadTypes";
+} from "./cwtTypes";
 import { CWTClaimsResult } from "./generalTypes";
 import { decodeCtiToJti } from "./jtiCti";
 import { currentTimestamp } from "./util";
